Extracted the tooltip data logic into hooks for reuse on the new design
