+++ ui/packages/shared/profile/src/GraphTooltipArrow/Content.tsx
@@ -18,27 +18,13 @@ import cx from 'classnames';
 import {CopyToClipboard} from 'react-copy-to-clipboard';
 import {Tooltip} from 'react-tooltip';
 
-import {QueryRequest_ReportType} from '@parca/client';
-import {Button, useParcaContext, useURLState} from '@parca/components';
-import {divide, getLastItem, valueFormatter, type NavigateFunction} from '@parca/utilities';
+import {Button, useParcaContext} from '@parca/components';
+import {getLastItem, type NavigateFunction} from '@parca/utilities';
 
-import {
-  FIELD_CUMULATIVE,
-  FIELD_DIFF,
-  FIELD_FUNCTION_FILE_NAME,
-  FIELD_FUNCTION_START_LINE,
-  FIELD_LABELS,
-  FIELD_LOCATION_ADDRESS,
-  FIELD_LOCATION_LINE,
-  FIELD_MAPPING_BUILD_ID,
-  FIELD_MAPPING_FILE,
-} from '../ProfileIcicleGraph/IcicleGraphArrow';
-import {nodeLabel} from '../ProfileIcicleGraph/IcicleGraphArrow/utils';
-import {ProfileSource} from '../ProfileSource';
-import {useProfileViewContext} from '../ProfileView/ProfileViewContext';
-import {useQuery} from '../useQuery';
 import {hexifyAddress, truncateString, truncateStringReverse} from '../utils';
 import {ExpandOnHover} from './ExpandOnHoverValue';
+import {useGraphTooltip} from './useGraphTooltip';
+import {useGraphTooltipMetaInfo} from './useGraphTooltipMetaInfo';
 
 let timeoutHandle: ReturnType<typeof setTimeout> | null = null;
 
@@ -69,13 +55,18 @@ const GraphTooltipArrowContent = ({
 }: GraphTooltipArrowContentProps): React.JSX.Element => {
   const [isCopied, setIsCopied] = useState<boolean>(false);
 
-  if (row === null) {
+  const graphTooltipData = useGraphTooltip({
+    table,
+    unit,
+    total,
+    totalUnfiltered,
+    row,
+    level,
+  });
+
+  if (graphTooltipData === null) {
     return <></>;
   }
-
-  const locationAddress: bigint = table.getChild(FIELD_LOCATION_ADDRESS)?.get(row) ?? 0n;
-  const cumulative: bigint = table.getChild(FIELD_CUMULATIVE)?.get(row) ?? 0n;
-  const diff: bigint = table.getChild(FIELD_DIFF)?.get(row) ?? 0n;
 
   const onCopy = (): void => {
     setIsCopied(true);
@@ -86,23 +77,7 @@ const GraphTooltipArrowContent = ({
     timeoutHandle = setTimeout(() => setIsCopied(false), 3000);
   };
 
-  const prevValue = cumulative - diff;
-  const diffRatio = diff !== 0n ? divide(diff, prevValue) : 0;
-  const diffSign = diff > 0 ? '+' : '';
-  const diffValueText = diffSign + valueFormatter(diff, unit, 1);
-  const diffPercentageText = diffSign + (diffRatio * 100).toFixed(2) + '%';
-  const diffText = `${diffValueText} (${diffPercentageText})`;
-
-  const name = nodeLabel(table, row, level, false);
-
-  const getTextForCumulative = (hoveringNodeCumulative: bigint): string => {
-    const filtered =
-      totalUnfiltered > total
-        ? ` / ${(100 * divide(hoveringNodeCumulative, total)).toFixed(2)}% of filtered`
-        : '';
-    return `${valueFormatter(hoveringNodeCumulative, unit, 2)}
-    (${(100 * divide(hoveringNodeCumulative, totalUnfiltered)).toFixed(2)}%${filtered})`;
-  };
+  const {name, locationAddress, cumulativeText, diffText, diff, row: rowNo} = graphTooltipData;
 
   return (
     <div className={`flex text-sm ${isFixed ? 'w-full' : ''}`}>
@@ -141,10 +116,8 @@ const GraphTooltipArrowContent = ({
                     <td className="w-1/4">Cumulative</td>
 
                     <td className="w-3/4">
-                      <CopyToClipboard onCopy={onCopy} text={getTextForCumulative(cumulative)}>
-                        <button className="cursor-pointer">
-                          {getTextForCumulative(cumulative)}
-                        </button>
+                      <CopyToClipboard onCopy={onCopy} text={cumulativeText}>
+                        <button className="cursor-pointer">{cumulativeText}</button>
                       </CopyToClipboard>
                     </td>
                   </tr>
@@ -160,7 +133,7 @@ const GraphTooltipArrowContent = ({
                   )}
                   <TooltipMetaInfo
                     table={table}
-                    row={row}
+                    row={rowNo}
                     onCopy={onCopy}
                     navigateTo={navigateTo}
                   />
@@ -190,97 +163,28 @@ const TooltipMetaInfo = ({
   onCopy: () => void;
   navigateTo: NavigateFunction;
 }): React.JSX.Element => {
-  const mappingFile: string = table.getChild(FIELD_MAPPING_FILE)?.get(row) ?? '';
-  const mappingBuildID: string = table.getChild(FIELD_MAPPING_BUILD_ID)?.get(row) ?? '';
-  const locationAddress: bigint = table.getChild(FIELD_LOCATION_ADDRESS)?.get(row) ?? 0n;
-  const locationLine: bigint = table.getChild(FIELD_LOCATION_LINE)?.get(row) ?? 0n;
-  const functionFilename: string = table.getChild(FIELD_FUNCTION_FILE_NAME)?.get(row) ?? '';
-  const functionStartLine: bigint = table.getChild(FIELD_FUNCTION_START_LINE)?.get(row) ?? 0n;
-  const lineNumber =
-    locationLine !== 0n ? locationLine : functionStartLine !== 0n ? functionStartLine : undefined;
-  const pprofLabelPrefix = 'pprof_labels.';
-  const labelColumnNames = table.schema.fields.filter(field =>
-    field.name.startsWith(pprofLabelPrefix)
-  );
+  const {
+    labelPairs,
+    functionFilename,
+    file,
+    openFile,
+    isSourceAvailable,
+    locationAddress,
+    mappingFile,
+    mappingBuildID,
+  } = useGraphTooltipMetaInfo({table, row, navigateTo});
+  const {enableSourcesView} = useParcaContext();
 
-  const {queryServiceClient, enableSourcesView} = useParcaContext();
-  const {profileSource} = useProfileViewContext();
-
-  const {isLoading: sourceLoading, response: sourceResponse} = useQuery(
-    queryServiceClient,
-    profileSource as ProfileSource,
-    QueryRequest_ReportType.SOURCE,
-    {
-      skip:
-        enableSourcesView === false ||
-        profileSource === undefined ||
-        // eslint-disable-next-line no-extra-boolean-cast
-        !Boolean(mappingBuildID) ||
-        // eslint-disable-next-line no-extra-boolean-cast
-        !Boolean(functionFilename),
-      sourceBuildID: mappingBuildID,
-      sourceFilename: functionFilename,
-      sourceOnly: true,
-    }
-  );
-
-  const isSourceAvailable = !sourceLoading && sourceResponse?.report != null;
-
-  const getTextForFile = (): string => {
-    if (functionFilename === '') return '<unknown>';
-
-    return `${functionFilename} ${lineNumber !== undefined ? ` +${lineNumber.toString()}` : ''}`;
-  };
-  const file = getTextForFile();
-
-  const labelPairs = labelColumnNames
-    .map((field, i) => [
-      labelColumnNames[i].name.slice(pprofLabelPrefix.length),
-      table.getChild(field.name)?.get(row) ?? '',
-    ])
-    .filter(value => value[1] !== '');
   const labels = labelPairs.map(
     (l): React.JSX.Element => (
       <span
         key={l[0]}
         className="mr-3 inline-block rounded-lg bg-gray-200 px-2 py-1 text-xs font-bold text-gray-700 dark:bg-gray-700 dark:text-gray-400"
       >
-        {`${l[0] as string}="${l[1] as string}"`}
+        {`${l[0]}="${l[1]}"`}
       </span>
     )
   );
-
-  const [dashboardItems, setDashboardItems] = useURLState({
-    param: 'dashboard_items',
-    navigateTo,
-  });
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [unusedBuildId, setSourceBuildId] = useURLState({
-    param: 'source_buildid',
-    navigateTo,
-  });
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [unusedFilename, setSourceFilename] = useURLState({
-    param: 'source_filename',
-    navigateTo,
-  });
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [unusedLine, setSourceLine] = useURLState({
-    param: 'source_line',
-    navigateTo,
-  });
-
-  const openFile = (): void => {
-    setDashboardItems([dashboardItems[0], 'source']);
-    setSourceBuildId(mappingBuildID);
-    setSourceFilename(functionFilename);
-    if (lineNumber !== undefined) {
-      setSourceLine(lineNumber.toString());
-    }
-  };
 
   return (
     <>
@@ -331,7 +235,7 @@ const TooltipMetaInfo = ({
       <tr>
         <td className="w-1/4">Binary</td>
         <td className="w-3/4 break-all">
-          {mappingFile === '' ? (
+          {mappingFile === null ? (
             <NoData />
           ) : (
             <CopyToClipboard onCopy={onCopy} text={mappingFile}>
@@ -343,7 +247,7 @@ const TooltipMetaInfo = ({
       <tr>
         <td className="w-1/4">Build Id</td>
         <td className="w-3/4 break-all">
-          {mappingBuildID === '' ? (
+          {mappingBuildID === null ? (
             <NoData />
           ) : (
             <CopyToClipboard onCopy={onCopy} text={mappingBuildID}>
